+++ tests/test_burst.py
@@ -2,7 +2,6 @@
 from collections import namedtuple
 from datetime import datetime
 from pathlib import Path
-from re import match
 from unittest.mock import patch
 
 import asf_search
@@ -107,20 +106,6 @@ def test_get_region_of_interest(tmp_path, orbit):
     assert not roi.intersects(ref_bbox_post)
 
 
-def test_get_product_name():
-    reference_name = 'S1_136231_IW2_20200604T022312_VV_7C85-BURST'
-    secondary_name = 'S1_136231_IW2_20200616T022313_VV_5D11-BURST'
-
-    name_20m = burst.get_product_name(reference_name, secondary_name, pixel_spacing=20.0)
-    name_80m = burst.get_product_name(reference_name, secondary_name, pixel_spacing=80)
-
-    assert match('[A-F0-9]{4}', name_20m[-4:]) is not None
-    assert match('[A-F0-9]{4}', name_80m[-4:]) is not None
-
-    assert name_20m.startswith('S1_136231_IW2_20200604_20200616_VV_INT20')
-    assert name_80m.startswith('S1_136231_IW2_20200604_20200616_VV_INT80')
-
-
 def mock_asf_search_results(
     slc_name: str, subswath: str, polarization: str, burst_index: int
 ) -> asf_search.ASFSearchResults:
@@ -188,18 +173,44 @@ def test_get_burst_params_multiple_results():
 
 
 def test_validate_bursts():
-    burst.validate_bursts('S1_030349_IW1_20230808T171601_VV_4A37-BURST', 'S1_030349_IW1_20230820T171602_VV_5AC3-BURST')
-    with pytest.raises(ValueError, match=r'.*polarizations are not the same.*'):
+    burst.validate_bursts('S1_000000_IW1_20200101T000000_VV_0000-BURST', 'S1_000000_IW1_20200201T000000_VV_0000-BURST')
+    burst.validate_bursts(
+        ['S1_000000_IW1_20200101T000000_VV_0000-BURST', 'S1_000001_IW1_20200101T000001_VV_0000-BURST'],
+        ['S1_000000_IW1_20200201T000000_VV_0000-BURST', 'S1_000001_IW1_20200201T000001_VV_0000-BURST'],
+    )
+
+    with pytest.raises(ValueError, match=r'Must include at least 1.*'):
+        burst.validate_bursts(['a'], [])
+
+    with pytest.raises(ValueError, match=r'Must have the same number.*'):
+        burst.validate_bursts(['a', 'b'], ['c'])
+
+    with pytest.raises(ValueError, match=r'.*burst ID sets do not match.*'):
         burst.validate_bursts(
-            'S1_215032_IW2_20230802T144608_VV_7EE2-BURST', 'S1_215032_IW2_20230721T144607_HH_B3FA-BURST'
+            ['S1_000000_IW1_20200101T000000_VV_0000-BURST'], ['S1_000000_IW2_20200201T000000_VV_0000-BURST']
         )
-    with pytest.raises(ValueError, match=r'.*burst IDs are not the same.*'):
+
+    with pytest.raises(ValueError, match=r'.*must have a single polarization.*'):
         burst.validate_bursts(
-            'S1_030349_IW1_20230808T171601_VV_4A37-BURST', 'S1_030348_IW1_20230820T171602_VV_5AC3-BURST'
+            ['S1_000000_IW1_20200101T000000_VV_0000-BURST', 'S1_000000_IW1_20200101T000000_VH_0000-BURST'],
+            ['S1_000000_IW1_20200201T000000_VV_0000-BURST', 'S1_000000_IW1_20200201T000000_VH_0000-BURST'],
         )
-    with pytest.raises(ValueError, match=r'.*only VV and HH.*'):
+
+    with pytest.raises(ValueError, match=r'.*polarization is not currently supported.*'):
         burst.validate_bursts(
-            'S1_030349_IW1_20230808T171601_VH_4A37-BURST', 'S1_030349_IW1_20230820T171602_VH_5AC3-BURST'
+            ['S1_000000_IW1_20200101T000000_VH_0000-BURST', 'S1_000000_IW1_20200101T000000_VH_0000-BURST'],
+            ['S1_000000_IW1_20200201T000000_VH_0000-BURST', 'S1_000000_IW1_20200201T000000_VH_0000-BURST'],
+        )
+
+    with pytest.raises(ValueError, match=r'.*must be from one date.*'):
+        burst.validate_bursts(
+            ['S1_000000_IW1_20200101T000000_VV_0000-BURST', 'S1_000001_IW1_20200101T000000_VV_0000-BURST'],
+            ['S1_000000_IW1_20200201T000000_VV_0000-BURST', 'S1_000001_IW1_20200202T000000_VV_0000-BURST'],
+        )
+
+    with pytest.raises(ValueError, match=r'Reference granules must be older.*'):
+        burst.validate_bursts(
+            'S1_000000_IW1_20200201T000000_VV_0000-BURST', 'S1_000000_IW1_20200101T000000_VV_0000-BURST'
         )