+++ tests/test_utils.py
@@ -4,12 +4,11 @@ import shutil
 from pathlib import Path
 from unittest.mock import patch
 
+import isceobj  # noqa
 import numpy as np
 from osgeo import gdal
 
 import hyp3_isce2.utils as utils
-
-import isceobj  # noqa
 
 
 gdal.UseExceptions()
@@ -52,13 +51,6 @@ def test_gdal_config_manager():
     assert gdal.GetConfigOption('OPTION2') == 'VALUE2'
     assert gdal.GetConfigOption('OPTION3') is None
     assert gdal.GetConfigOption('OPTION4') == 'VALUE4'
-
-
-def test_oldest_granule_first():
-    oldest = 'S1_249434_IW1_20230511T170732_VV_07DE-BURST'
-    latest = 'S1_249434_IW1_20230523T170733_VV_8850-BURST'
-    assert utils.oldest_granule_first(oldest, latest) == (oldest, latest)
-    assert utils.oldest_granule_first(latest, oldest) == (oldest, latest)
 
 
 def test_make_browse_image():